+++ app/[locale]/(dashboard)/contactUs/page.tsx
@@ -1,6 +1,6 @@
 import DashboardLayout from "../DashboardLayout";
-import ContactInfo from "@/components/contactUs/ContactInfo";
-import ContactBanners from "@/components/contactUs/ContactBanners";
+import ContactInfo from "../../../../components/contactUs/ContactInfo";
+import ContactBanners from "../../../../components/contactUs/ContactBanners";
 
 function Contacts() {
   return (
