+++ components/banner/Banner.tsx
