+++ components/header/Header.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import LegoLogo from "../icons/LEGO_logo.png";
 import Link from "next/link";
 import LogoutButton from "./LogoutButton";
-import { Logout } from "@/app/actions";
+import { Logout } from "../../app/actions";
 import ThemeSwitcher from "./ThemeSwitcher";
 import { getTranslations } from "next-intl/server";
 import LocaleSwitcher from "./LocaleSwitcher";
@@ -11,7 +11,7 @@ async function Header() {
   const t = await getTranslations("navigation")
   const handleLogout = async () => {
     "use server";
-    await Logout()
+    await Logout();
   }
 
   return (
@@ -37,13 +37,13 @@ async function Header() {
               <Link href="/contactUs">{t("contactus")}</Link>
             </li>
             <li className="uppercase font-bold text-base">
-             <ThemeSwitcher />
+              <ThemeSwitcher />
             </li>
             <li className="uppercase font-bold text-base">
               <LocaleSwitcher />
             </li>
             <li className="uppercase font-bold text-base">
-             <LogoutButton handleLogout={handleLogout} />
+              <LogoutButton handleLogout={handleLogout} />
             </li>
           </ul>
         </nav>
