+++ components/header/LocaleSwitcherSelect.tsx
@@ -1,26 +1,35 @@
 'use client';
 
 import clsx from 'clsx';
-import {useParams} from 'next/navigation';
-import {useTransition} from 'react';
-import {useRouter, usePathname} from '@/navigation';
+import { useParams } from 'next/navigation';
+import { useTransition, ReactNode, ChangeEvent } from 'react';
+import { useRouter, usePathname } from '../../navigation';
+
+type Props = {
+  children: ReactNode;
+  defaultValue: string;
+  label: string;
+};
 
 export default function LocaleSwitcherSelect({
   children,
   defaultValue,
   label
-}) {
+}: Props) {
   const router = useRouter();
   const [isPending, startTransition] = useTransition();
   const pathname = usePathname();
   const params = useParams();
 
-  function onSelectChange(event) {
+  function onSelectChange(event: ChangeEvent<HTMLSelectElement>) {
     const nextLocale = event.target.value;
     startTransition(() => {
       router.replace(
-        {pathname, params},
-        {locale: nextLocale}
+        // @ts-expect-error -- TypeScript will validate that only known `params`
+        // are used in combination with a given `pathname`. Since the two will
+        // always match for the current route, we can skip runtime checks.
+        { pathname, params },
+        { locale: nextLocale }
       );
     });
   }
