+++ package.json
@@ -14,18 +14,20 @@
     "@fortawesome/free-brands-svg-icons": "^6.5.2",
     "@fortawesome/free-solid-svg-icons": "^6.5.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@types/react": "^18.3.0",
+    "@types/react": "^18.3.1",
     "axios": "^1.6.8",
     "clsx": "^2.1.1",
     "negotiator": "^0.6.3",
     "next": "14.1.4",
     "next-intl": "^3.11.3",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "sharp": "^0.33.3"
   },
   "devDependencies": {
     "@types/node": "20.12.7",
     "autoprefixer": "^10.0.1",
+    "encoding": "^0.1.13",
     "eslint": "^8",
     "eslint-config-next": "14.1.4",
     "postcss": "^8",
